Add `FaceBuilder::triangle`
