+++ crates/fj-kernel/src/builder/face.rs
@@ -1,13 +1,14 @@
 use fj_interop::mesh::Color;
+use fj_math::Point;
 
 use crate::{
     insert::Insert,
-    objects::{Face, Objects, Surface},
+    objects::{Cycle, Face, GlobalEdge, Objects, Surface},
     services::Service,
     storage::Handle,
 };
 
-use super::CycleBuilder;
+use super::{CycleBuilder, HalfEdgeBuilder, SurfaceBuilder};
 
 /// Builder API for [`Face`]
 pub struct FaceBuilder {
@@ -25,6 +26,37 @@ impl FaceBuilder {
             interiors: Vec::new(),
             color: None,
         }
+    }
+
+    /// Create a triangle
+    pub fn triangle(
+        points: [impl Into<Point<3>>; 3],
+        objects: &mut Service<Objects>,
+    ) -> (Handle<Face>, [Handle<GlobalEdge>; 3]) {
+        let [a, b, c] = points.map(Into::into);
+
+        let surface =
+            SurfaceBuilder::plane_from_points([a, b, c]).insert(objects);
+        let (exterior, global_edges) = {
+            let half_edges = [[a, b], [b, c], [c, a]].map(|points| {
+                HalfEdgeBuilder::line_segment_from_global_points(
+                    points, &surface, None,
+                )
+                .build(objects)
+                .insert(objects)
+            });
+
+            let cycle = Cycle::new(half_edges.clone()).insert(objects);
+
+            let global_edges =
+                half_edges.map(|half_edge| half_edge.global_form().clone());
+
+            (cycle, global_edges)
+        };
+
+        let face = Face::new(surface, exterior, [], None).insert(objects);
+
+        (face, global_edges)
     }
 
     /// Replace the face's exterior cycle
